Moved shared functions to a shared file
The functions dl_internal_str_format and dl_grow_array were duplicated in multiple source files. I moved them to a new shared header-file, dl_util.h in order to be able to just include all the source-files to one unity-build.
